Only have main task write out model flags
Before this commit, if the print_clubb_config_flags was called all tasks would
write out the 39 listed flags. For simulations with significantly high task
counts, this would cause an enormous amount of output to be generated. For
instance, 300 fully nodes on Cheyenne would cause 421,200 lines of output.

This commit reduces the number of output to 39 messages.